Update repository with NeurIPS submission

+++ scripts/exact_gp_comparison.py
@@ -108,7 +108,7 @@ layers = init_layers(train_dataset.inputs, train_dataset.output_dim, **vars(args
 
 
 # Instantiate Likelihood
-ll = args.likelihood(device=args.device, dtype = args.dtype)
+ll = args.likelihood(device=args.device, dtype=args.dtype)
 
 dvip = DVIP_Base(
     ll,
